Add types, loading from files, and meson build
Defined type Matrix2D as a vector of vector to be able to get data
from files without having to hardcode the dimensions like would be
needed if using standard arrays. Later on, in order to copy to gpu
memory, we will know the size from the Matrix2D.

Currently errors are just being sent to the std error output, but
later (no much later) will be advisable to have a logging system
set.

Also have to figure out a better way of doing unit tests taking
advantage of the `test` function of meson.