Add interpolation functions
Added bilinear interpolations using four and six point formulae. For
the six point, falls back to four at the borders.

Would be better to have a type that encopsulates the matrix of the
function and also the domain that it maps. Something like:

struct Field{
	double* f;
	int Nx, Ny;
	double x_min, x_max, y_min, y_max;

	double operator()(double x, double y){
		return some_interpolation_func(x, y);
	}
}

But I don't know if that would work with CUDA.