Add Chebyshev polynomial expansion
Inside the `#if defined(COMPILE_CUDA)` will later be the kernels to
be compiled with nvcc. All the other functions outside the `#ifdef`
should be both compatible with C++ and CUDA devices. However, inside
those functions there are C++ style outputs, that will not run with
CUDA as fas as I know. Need to change the `std::cerr` inside functions
that will be called from devices.