Add script to copy files during build
`copy.py` copies the file given in the first argument to the file given
as second argument, from the project source directory inside the subdir
`meson.build` is, to the equivalent subdir inside the build directory.

This will be good for having input files and expected output, during
build these could be copied in order to test via `meson test`.