Add Chebyshev expansion and evaluation
There are some fixes included in this commit. The order of the x and y
indexes was fixed so that the first index always corresponds to x. Also
fixed th Chebyshev coefficients, that needed normalization.

Right now the testing is based on visual comparison with previous
results, this should be improved to an automatic test. The output
of the Chebyshev expansion "looks" very close to the original values.
This gets worse for large (n = 100) or low (n = 10) order of the
expansion, for the large it seems to over-specialize on the discrete
values the fitting function takes and the image gets blocky.