Create pair and matrix2D types and modify files IO
CUDA supports C++ templates and classes, unfortunately it does not
support the STL, so we'll need to create our own replacements suited
for our needs. I created the `pair` and `matrix2D` types that will most
probably be tweaked as the project goes on.

I modified the files IO in order to be compatible with the new types
and also put the restriction that matrix files should have the shape
indicated in them. The files format might (and most probably will) be
change later on.

With `load` I would really like to be able to just return the matrix
and not really on changing the shape of one passed by reference. It
would be more semantic that way. But first would need to make sure that
the compiler will not copy the entire matrix.

For meson I have to figure out how to check that the files test gave
the same output as the input. `cmp` command of linux can verify this,
just have to learn how to automate it in meson.

When the types are tested in plain C++ will add the CUDA __device__
wherever needed to make the code run on GPU.