Add Chebyshev expansion tests
The Chebyshev expansion was failing because of the error inside
interpolations that was solved by commit e932c01. I added a test to
Chebyshev expansion that also requires interpolations to be working
properly. However, the results with the expansion are not very good
the accuracy is about 0.02 for order 23 and doesn't increase with
higher orders, in fact, the best result comparing element by element
is obtained for order 13. This might have to do with the fact that
we are interpolating first in order to make the expansion.