Add second kind Chebyshev polynomials
The derivative of the first kind Chebyshev polynomials is very concisely
expressed in terms of the second kind Chebyshev polynomials without the
need of considering x = -1 and x = 1 cases apart.

The second kind polynomials can also be calculated using a recurrence
formula, hence there is no trade-off it's a plain improvement.