Fix files read and dump format
A silly mistake from the very beginning, the incongruence with the
results from cclauser where because of this. Still need to prove if the
derivative from an expansion of Chebyshev works with a regular function.

I added the plot_b_field.py with which I visually compare the results.

There is still a difference regarding the direction of the magnetic
field. The formula in this repository is the correct one, but maybe I'm
ignoring some assumptions about the original efit file.