Add move constructors and read first Lines
Added the part of code to read the first line of geqdsk file.

In order for the returned Equilibrium to be moved in return I
implemented the move constructors for the Array and Matrix2D classes.
This allows objects of those classes to be returned from functions
painlessly swapping the pointers and keeping the data on the heap.
This way we can use return statement which is far more elegant than
passing out parameters.