Add Ordinary Differential Equations INTegrator (odeint)
The odeint section is meant to be as modular as possible, trying to
follow the style of boost/numeric/odeint but keeping it at the minimum
necessary. The scheme is as follow:

An odeint has the form sys(x, dxdt, t), where sys represents the
equation system that relates the variable x with its derivative with
respect to the variable t (dxdt) and t. So the odeint is represented by
a function or a callable object in the form

`void sys(const State& x, State& dxdt, Scalar t);`

where dxdt is an out argument.

An integration method generally consist in a sequence of steps repeated
until the desired time is reached. So the function `integrate` takes a
system and a stepper and calculates steps in fixed size time deltas (dt)
from t0 to Nsteps * dt. This function may be called with an observer
in order to observe the evolution of the state during the integration.

The integration method is implemented in the stepper object that the
integration function takes as argument. The steppers must have a method
do_step(System sys, State x, Scalar t, Scalar dt) that makes x evolve
from t to t + dt.

In this commit I added the EulerStepper just for testing purposes; more
accurate methods of integration should be implemented.