Add arbitrary fixed size Vector
Templated the Vector class to have an arbitrary number of components
via template, this way 3D vectors and state (position + velocity) can
be just special cases.

The constructor of the general size Vector has problems casting to
double when numbers are resented like integers, so we'll have to keep
it in mind if we set vector by hand.

I'll later try to use std::initializer_list, but I'm not sure if that
will work with CUDA.