Add Magnetic Field calculation for evolution
Every step the function interpolates the magnetic field from the
matrices refined via Chebyshev expansion.

Fixed the calculation of the toroidal magnetic field.

I'm not very happy with the results from test/magnetic_field.cpp. I
think there is a problem with the units or the dimensionless factors,
or maybe the values I'm putting in are not correct.