Fix magnetic fields calculations
Previous versions weren't working because of the limits, I think the
interpolation function was being called out of bounds. I "fixed" it
using the boundary limits as the limit to the chebyshev calculation.
It could not work in some cases because it seems that the G-EQDSK files
can lack the boundaries descriptions.

TODO: Fix the chebyshev/interpolation so that it can handle the job
just a pair of limits.