Add documentation
The documentation of the classes and functions inside the code is
generated with Doxygen; it can later be accessed with Sphinx + Breathe
in order to build a more enriched documentation. I linked the repo
to Read the Docs, that builds a new version of the documentations every
time it is updated and hosts the documentation website. Hopefully it
works.