Change `ParticleSpecies` from templated to base
class so we use polymorphism here. This could be useful if some of the
particle species implement their temperature and density profiles in
diferent fashions.

I find it kind of ugly to have to use pointers to the class so that the
polymorphism works but ¯\_(ツ)_/¯

Also, I'm not sure about the virtual destructor and if it is doing what
it should or I am just getting rid of the warning without solving the
actual issue.