Add CUDA to collisionless integrate
Right now there is a long waring message when cimpiling with nvcc the
test integrate.cu. This si because the integrate fucntion is __host__
__device__ and is being called from a __host__ and then calls something
that is __host__. NVCC doesn't check that I'm not intending on calling
the __host__ std::cout from from the __device__ part of integrate. This
might maybe lead to errors later, so the next step would be to make all
observers Array based or leave integrate __device__ only and compile
with g++ if I want a host version.