Fix memory leakage in device
Previously the copy from host to device of the containers in Equilibrium
and MagneticFieldMatrix were being made inside the copy constructor.
This was not ok as in each call to the kernel the copy contructor would
allocate new space in the device.

The solution was something that I was thinking of for some time but just
for coherence and not because I knew it could solve a problem:

*make every object that needs to copy a container to the device have a
`construct_in_host_for_device` method*