Add Boris stepper for integration
The Boris method is the "de facto" method for integrating solving the
motion of charged particles subjected to the Lorentz force. I tested
the method in CPU and it it faster than RK46NL with -O0, -O1 and -O2;
for -O3 somehow RK46NL performed better. Performance un GPU is yet to
be tested.