Add out of bound error to Array and Matrix2D
It was absolutely nonseinsical the way I emplemented Array and Matrix2D
to never fail. This lead to many error hiding at plain sight and many
ours/days of frustration. From now on this will not be the case anymore
(at least not because of this).

I added some fail tests too